libgcc may be in /usr/local/lib

+++ config/software/libgcc.rb
@@ -2,5 +2,9 @@ name "libgcc"
 description "On UNIX systems where we bootstrap a compiler, copy the libgcc"
 
 build do
-  command "cp /usr/lib/libgcc_s.so.1 #{install_dir}/embedded/lib/"
+  if File.exists?("/usr/lib/libgcc_s.so.1")
+    command "cp /usr/lib/libgcc_s.so.1 #{install_dir}/embedded/lib/"
+  else
+    command "cp /usr/local/lib/libgcc_s.so.1 #{install_dir}/embedded/lib/"
+  end
 end
